Have pngpack-py compile libpngpack with it
There were issues with loading libpngpack as a dynamic library,
so instead of that, compile pngpack.c link libpng directly.

The downside is that the compilation is split up into 2 parts now,
and dependencies of libpngpack need to be maintained in 2 locations:
in libpngpack's CMakeFile and in pngpack-py's setup file.

Compiling libpngpack as a static library does not fix this issue,
as libpng still needs to be linked to the final dynamic library
one way or another. Attempts were made to merge libpngpack and libpng
static library files but I have not had success (yet?).